Move broadening functions into spectra
